Add MacOS dist file
