--- RPGDictionary.py
@@ -1,2 +1,0 @@
-# Copyright 2018 Ursula Searle
-rpgsystems ={'7c': {'name': '7c:7th Sea', 'sides': 10, 'explodes': 'Y'}}
