+++ src/RPGDictionary.py
@@ -1,0 +1,2 @@
+# Copyright 2018 Ursula Searle
+rpgsystems ={'7c': {'name': '7c:7th Sea', 'sides': 10, 'explodes': 'Y'}}
